backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# LLM & RAG
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.5
groq==0.9.0
sentence-transformers==3.0.1
torch==2.3.1
tiktoken==0.7.0

# Vector store
qdrant-client==1.9.1

# Graph database
neo4j==5.21.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Job queue
celery==5.4.0
redis==5.0.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# Document processing
pypdf==4.2.0
python-docx==1.1.2
pytube==15.0.0

# HTTP & utilities
httpx==0.27.0
aiofiles==23.2.1
python-slugify==8.0.4
structlog==24.2.0
